requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
pytesseract
opencv-python
pandas
pillow
pywin32
